Issue #GH-381: add param to turn off case flipping in TextCorpus
If a corpus is known to be mono-lingual, then it is useful to turn off
the default behavior of randomly swapping the case of a small
percentage of lines.

This fix adds a top level parameter that by default preserves the
original behavior. The user can set this parameter to False if he or
she does not want the data randomly swapped.