Show allennlp import errors when installed
When allennlp is installed, but there is an issue with its installation (for example, the version of `protobuf` installed is too old), Flair produces the unhelpful error message "The library "allennlp" is not installed! To use ELMoEmbeddings, please first install with "pip install allennlp". These changes limit the scope of errors caught when loading `allennlp` in `embeddings.py` to only include `ModuleNotFoundError`s. This allows errors in actually loading the installed `allennlp` module to be displayed to the user.