Change way of returning predictions in the flair eval routine.
Problem in the new way of returning predictions: the predictions are only contained in sentences if it was chosen memory_mode="full" during creation of the ClassificationDataset in the ClassificationCorpus creation. If the ClassificationCorpus has memory mode "partial", then the predicted labels are not contained in sentences in any case so the following optional removal has no effect. Predictions won't be accessible outside the eval routine in this case regardless whether return_predictions is True or False. TODO: fix this